feat: 당첨시 수익률 음수로 출력되는 현상 수정
- 당첨금액 누적합계 변수의 타입을 int -> long로 변경.

+++ src/nb993/controller/LottoController.java
@@ -73,7 +73,7 @@ public class LottoController {
         for (Map.Entry<Rank, Integer> entry : rankResult.entrySet()) {
             Rank rank = entry.getKey();
             int count = entry.getValue();
-            winningMoney += rank.getWinningMoney() * count;
+            winningMoney += (long)rank.getWinningMoney() * count;
         }
         return winningMoney;
     }
